// app/modules/shared/components/chart-widget/constants/chart-element-colors.enum.ts
// refers to client/laCommon/styles/less/atomic-stages/atoms/__variable/_color/atoms__variable_color.less
export enum ChartElementColors {
  REGENT_GRAY_BASE = '#7D889B',
  GEYSER_BASE = '#D6DCE5',
  OXFORD_BLUE_BASE = '#354053',
  OTHER_SERIES = '#F3D49C',
  APPLE_BASE = '#31b937',
  RED_ORANGE_BASE = '#FF3635',
  RED_ORANGE_BASE_FADE = '#ff363560',
}